--- a/src/NSubstitute/Core/CallInfoFactory.cs
+++ b/src/NSubstitute/Core/CallInfoFactory.cs
@@ -14,12 +14,12 @@ namespace NSubstitute.Core
         private static IEnumerable<Argument> GetArgumentsFromCall(ICall call)
         {
             var values = call.GetArguments();
-            var types = call.GetParameterInfos().Select(x => x.ParameterType).ToArray();
+            var parameterInfos = call.GetParameterInfos();
 
             for (var index = 0; index < values.Length; index++)
             {
                 var i = index;
-                yield return new Argument(types[i], () => values[i], x => values[i] = x);
+                yield return new Argument(parameterInfos[i].ParameterType, () => values[i], x => values[i] = x);
             }
         }
     }