--- a/src/NSubstitute/Routing/Route.cs
+++ b/src/NSubstitute/Routing/Route.cs
@@ -1,35 +1,40 @@
 using System.Collections.Generic;
-using System.Linq;
 using NSubstitute.Core;
 
 namespace NSubstitute.Routing
 {
-    public enum RouteType { Other, RecordReplay };
+    public enum RouteType { Other, RecordReplay }
 
     public class Route : IRoute
     {
-        private readonly bool _isRecordReplayRoute;
-        private readonly IEnumerable<ICallHandler> _handlers;
+        private readonly ICallHandler[] _handlers;
 
-        public Route(IEnumerable<ICallHandler> handlers) : this(RouteType.Other, handlers) { }
+        public Route(ICallHandler[] handlers) : this(RouteType.Other, handlers) { }
 
-        public Route(RouteType routeType, IEnumerable<ICallHandler> handlers)
+        public Route(RouteType routeType, ICallHandler[] handlers)
         {
-            _isRecordReplayRoute = routeType == RouteType.RecordReplay;
+            IsRecordReplayRoute = routeType == RouteType.RecordReplay;
             _handlers = handlers;
         }
 
-        public bool IsRecordReplayRoute { get { return _isRecordReplayRoute; } }
+        public bool IsRecordReplayRoute { get; }
 
-        public IEnumerable<ICallHandler> Handlers
-        {
-            get { return _handlers; }
-        }
+        public IEnumerable<ICallHandler> Handlers => _handlers;
 
         public object Handle(ICall call)
         {
-            var result = Handlers.Select(x => x.Handle(call)).FirstOrDefault(x => x.HasReturnValue);
-            return result == null ? null : result.ReturnValue;
+            // This is a hot method which is invoked frequently and has major impact on performance.
+            // Therefore, the LINQ cycle was unwinded to for loop.
+            for (int i = 0; i < _handlers.Length; i++)
+            {
+                var result = _handlers[i].Handle(call);
+                if (result.HasReturnValue)
+                {
+                    return result.ReturnValue;
+                }
+            }
+
+            return null;
         }
     }
 }