I benchmarked the code and found a couple of hot places that could improve call dispatch performance a bit. Also change inside the `Router.Route()` method allows to easily debug code and makes call stack nicer.

Benchmark results:
```
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3914066 Hz, Resolution=255.4888 ns, Timer=TSC
  [Host]     : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  DefaultJob : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0

BASE:
                            Method |        Mean |      Error |     StdDev |
---------------------------------- |------------:|-----------:|-----------:|
 InvokeRegularMethodInterfaceProxy | 3,974.54 ns | 48.6592 ns | 45.5159 ns |
     InvokeRegularMethodClassProxy | 6,180.20 ns | 39.3871 ns | 34.9156 ns |

     

FIRST COMMIT:
                            Method |     Mean |     Error |    StdDev |
---------------------------------- |---------:|----------:|----------:|
 InvokeRegularMethodInterfaceProxy | 3.335 us | 0.0315 us | 0.0294 us |
     InvokeRegularMethodClassProxy | 5.636 us | 0.0590 us | 0.0552 us |
     
     
SECOND:
                            Method |     Mean |     Error |    StdDev |
---------------------------------- |---------:|----------:|----------:|
 InvokeRegularMethodInterfaceProxy | 3.236 us | 0.0643 us | 0.0601 us |
     InvokeRegularMethodClassProxy | 5.353 us | 0.1326 us | 0.1176 us |
```

@dtchepak Please take a look.